// types.ts
export interface ClassType {
  id: string;
  title: string;
  description: string;
  image: string;
  videoUrl: string;
}

export interface Amenity {
  id: string;
  title: string;
  description: string;
  image: string;
}

export interface Trainer {
  id: string;
  name: string;
  specialty: string;
  image: string;
}